app/flask/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
apispec==5.0.0
apispec-webframeworks==0.5.2
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
Flask==2.0.1
Flask-Admin==1.5.8
flask-apispec==0.11.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
webargs==8.0.0
Werkzeug==2.0.1
WTForms==2.3.3
